persistence/src/test/java/gov/nasa/arc/mct/persistence/util/HibernateUtilTest.java
/*******************************************************************************
 * Mission Control Technologies, Copyright (c) 2009-2012, United States Government
 * as represented by the Administrator of the National Aeronautics and Space 
 * Administration. All rights reserved.
 *
 * The MCT platform is licensed under the Apache License, Version 2.0 (the 
 * "License"); you may not use this file except in compliance with the License. 
 * You may obtain a copy of the License at 
 * http://www.apache.org/licenses/LICENSE-2.0.
 *
 * Unless required by applicable law or agreed to in writing, software 
 * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT 
 * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the 
 * License for the specific language governing permissions and limitations under 
 * the License.
 *
 * MCT includes source code licensed under additional open source licenses. See 
 * the MCT Open Source Licenses file included with this distribution or the About 
 * MCT Licenses dialog available at runtime from the MCT Help menu for additional 
 * information. 
 *******************************************************************************/
package gov.nasa.arc.mct.persistence.util;
/**
 * HibernateRetrievalTest.java Sep 28, 2008
 * 
 * This code is property of the National Aeronautics and Space Administration
 * and was produced for the Mission Control Technologies (MCT) Project.
 * 
 */

import java.io.IOException;

import org.testng.Assert;
import org.testng.annotations.AfterTest;
import org.testng.annotations.BeforeTest;
import org.testng.annotations.Test;

public class HibernateUtilTest {

    @BeforeTest
    public void setUp() throws IOException {
        HibernateUtil.initSessionFactory("/hibernate_derby_test.cfg.xml");
    }

    @AfterTest
    public void shutdown() {
    }

    @Test
    public void getterTest() {
       Assert.assertNotNull(HibernateUtil.getDatabaseUserName());
       Assert.assertNotNull(HibernateUtil.getDatabaseName()); 
       Assert.assertTrue(HibernateUtil.getDatabaseURL().startsWith("jdbc:mysql"));   
    }
}
